+++ pixelfed/CHANGELOG.md
@@ -8,3 +8,5 @@
 * Redis must be running before pixelfed is configured, optionally user www must be part of redis group
 * Add www user to redis group
 * Remove redis socket in favour of remote redis jail over tcp
+* Postgresql databases aren't automatically created, we need a step to do this
+* Remove horizon:publish step as no longer needed